Rename: `Modal`컴포넌트 폴더 위치 변경
